Refactor and clean up pytest suite for logic verification
- Consolidated Noise Profiler logic tests into `tests/logic_verification/test_noise_profiler_widget_logic.py`.
- Moved Noise Profiler integration test to `tests/gui/test_noise_profiler_gui.py`.
- Removed redundant and broken Lock-in Amplifier tests (`test_lockin_stats.py`, `test_lockin_thd_callback.py`).
- Renamed `test_lockin_buffer.py` to `test_lockin_amplifier_logic.py` for clarity.
- Moved `test_frequency_counter_logic.py` and `test_allan_worker.py` from `functional` to `logic_verification`.
- Moved `test_analysis_worker.py` from `gui` to `logic_verification/test_advanced_distortion_worker.py`.
- Preserved critical measurement algorithm tests (`test_analysis_lockin.py`, `test_noise_profile_logic.py`).
