+++ tests/gui/test_noise_profiler_gui.py
