+++ tests/logic_verification/test_frequency_counter_logic.py
