+++ tests/logic_verification/test_lockin_amplifier_logic.py
