--- tests/logic_verification/test_lockin_stats.py
@@ -1,94 +1,0 @@
-
-import numpy as np
-from collections import deque
-from src.gui.widgets.lock_in_frequency_counter import LockInFrequencyCounter
-
-class MockAudioEngine:
-    def __init__(self):
-        self.sample_rate = 48000.0
-
-    def register_callback(self, cb):
-        return 1
-
-    def unregister_callback(self, id):
-        pass
-
-def test_nco_stats():
-    """Verify NCO statistics calculation."""
-    engine = MockAudioEngine()
-    counter = LockInFrequencyCounter(engine)
-
-    # Setup
-    counter.gen_frequency = 1000.0
-    counter.locked = True
-    counter.nco_avg_count = 5
-    counter.nco_history = deque(maxlen=5)
-
-    # Simulate processing loops with varying NCO frequencies
-    # We cheat and directly modify 'gen_frequency' and call the logic snippet
-    # Since we can't easily run the full process_data loop 5 times with exact outcomes 
-    # without robust mocking of the signal.
-    # Instead, let's verify the logic by spoofing the "lock" update part.
-
-    # However, process_data does the appending. So we should run process_data.
-
-    # To make process_data run without error and update stats, we need to provide input data
-    # that results in a successful "signal present" check.
-
-    # sr = engine.sample_rate
-    counter.buffer_size = 1024
-    counter.input_data = np.zeros((counter.buffer_size, 2))
-    counter.is_running = True
-    counter._estimates_discarded = 100 # Ready
-    counter.start_time = 1.0
-
-    # Create a clean signal that matches NCO, so Delta F is 0.
-    # But we WANT NCO to change.
-    # If locked, NCO changes based on PID.
-
-    # t = np.arange(counter.buffer_size) / sr
-
-    # Iter 1: Delta F = 1.0 -> NCO increases
-    counter.pid.kp = 1.0
-    counter.pid.ki = 0
-    counter.pid.kd = 0
-
-    # We want to force NCO to be specific values to check Mean/Std.
-    # Best way: Simulate the values that would be appended.
-
-    vals = [1000.0, 1001.0, 1002.0, 1003.0, 1004.0]
-
-    for v in vals:
-        # Manually invoke the stats logic or minimal reproduction
-        counter.gen_frequency = v
-        counter.nco_history.append(v)
-
-        # Calculate expected
-        data = list(counter.nco_history)
-        # expected_mean = np.mean(data)
-        # expected_std = np.std(data)
-
-        # In real code this happens inside process_data. 
-        # let's just assert our manual update matches numpy
-        pass
-
-    # Now verify the app's logic matches
-    # Update internal state as if process_data ran
-    data = list(counter.nco_history)
-    counter.nco_mean = float(np.mean(data))
-    counter.nco_std = float(np.std(data))
-
-    assert counter.nco_mean == 1002.0
-    assert abs(counter.nco_std - np.std(vals)) < 1e-6
-
-    # Test Maxlen
-    counter.gen_frequency = 1005.0
-    counter.nco_history.append(1005.0)
-    # [1001, 1002, 1003, 1004, 1005] -> Mean 1003
-
-    data = list(counter.nco_history)
-    counter.nco_mean = float(np.mean(data))
-
-    assert counter.nco_mean == 1003.0
-    assert len(counter.nco_history) == 5
-
